Homme: take exec space type decl out of ExecSpaceDefs.hpp
