+++ components/homme/src/share/cxx/ExecSpaceDefs.hpp
@@ -14,84 +14,14 @@
 
 #include <Kokkos_Core.hpp>
 
+#include "HommeExecSpace.hpp"
+
 #include "Config.hpp"
 #include "Dimensions.hpp"
 #include "vector/vector_pragmas.hpp"
 
 namespace Homme
 {
-
-// Some in-house names for Kokkos exec spaces, which are
-// always defined, possibly as alias of void
-
-#ifdef HOMMEXX_ENABLE_GPU
-
-#ifdef KOKKOS_ENABLE_CUDA
-using HommexxGPU = Kokkos::Cuda;
-#endif
-
-#ifdef KOKKOS_ENABLE_HIP
-using HommexxGPU = Kokkos::Experimental::HIP;
-#endif
-
-#ifdef KOKKOS_ENABLE_SYCL
-using HommexxGPU = Kokkos::Experimental::SYCL;
-#endif
-
-#else
-using HommexxGPU = void;
-#endif
-
-
-
-#ifdef KOKKOS_ENABLE_OPENMP
-using Hommexx_OpenMP = Kokkos::OpenMP;
-#else
-using Hommexx_OpenMP = void;
-#endif
-
-#ifdef KOKKOS_ENABLE_PTHREADS
-using Hommexx_Threads = Kokkos::Threads;
-#else
-using Hommexx_Threads = void;
-#endif
-
-#ifdef KOKKOS_ENABLE_SERIAL
-using Hommexx_Serial = Kokkos::Serial;
-#else
-using Hommexx_Serial = void;
-#endif
-
-#ifdef HOMMEXX_ENABLE_GPU
-# define HOMMEXX_STATIC
-#else
-# define HOMMEXX_STATIC static
-#endif
-
-// Selecting the execution space. If no specific request, use Kokkos default
-// exec space
-#ifdef HOMMEXX_ENABLE_GPU
-using ExecSpace = HommexxGPU;
-#elif defined(HOMMEXX_OPENMP_SPACE)
-using ExecSpace = Hommexx_OpenMP;
-#elif defined(HOMMEXX_THREADS_SPACE)
-using ExecSpace = Hommexx_Threads;
-#elif defined(HOMMEXX_SERIAL_SPACE)
-using ExecSpace = Hommexx_Serial;
-#elif defined(HOMMEXX_DEFAULT_SPACE)
-using ExecSpace = Kokkos::DefaultExecutionSpace::execution_space;
-#else
-#error "No valid execution space choice"
-#endif // HOMMEXX_EXEC_SPACE
-
-static_assert (!std::is_same<ExecSpace,void>::value,
-               "Error! You are trying to use an ExecutionSpace not enabled in Kokkos.\n");
-
-template <typename ExeSpace>
-struct OnGpu { enum : bool { value = false }; };
-
-template <>
-struct OnGpu<HommexxGPU> { enum : bool { value = true }; };
 
 // What follows provides utilities to parameterize the parallel machine (CPU/KNL
 // cores within a rank, GPU attached to a rank) optimally. The parameterization
@@ -199,18 +129,6 @@ static int get_num_concurrent_teams(const int num_parallel_iterations) {
   return get_num_concurrent_teams(
       get_default_team_policy<ExecSpace, Tags...>(num_parallel_iterations));
 }
-
-// A templated typedef for MD range policy (used in RK stages)
-template<typename ExecutionSpace, int Rank>
-using MDRangePolicy = Kokkos::MDRangePolicy
-                          < ExecutionSpace,
-                            Kokkos::Rank
-                              < Rank,
-                                Kokkos::Iterate::Right,
-                                Kokkos::Iterate::Right
-                              >,
-                            Kokkos::IndexType<int>
-                          >;
 
 template <typename ExeSpace>
 struct Memory {
